mjakubowski84#313 Add pekko support
removed unnecessary compat elements, general cleanup, documentation

fix scaladoc compilation
