+++ project/ActorLibCross.scala
@@ -1,0 +1,4 @@
+case class ActorLibCross(
+    override val idSuffix: String,
+    override val directorySuffix: String
+) extends sbt.VirtualAxis.WeakAxis
