+++ project/metals.sbt
@@ -2,5 +2,5 @@
 
 // This file enables sbt-bloop to create bloop config files.
 
-addSbtPlugin("ch.epfl.scala" % "sbt-bloop" % "1.5.8")
+addSbtPlugin("ch.epfl.scala" % "sbt-bloop" % "1.5.11")
