+++ site/docs/docs/examples.md
@@ -6,6 +6,6 @@ permalink: docs/examples/
 
 # Examples
 
-Please check [examples](https://github.com/mjakubowski84/parquet4s/blob/master/examples) where you can find simple code covering basics for `core`, `akka` and `fs2` modules.
+Please check [examples](https://github.com/mjakubowski84/parquet4s/blob/master/examples) where you can find simple code covering basics for `core`, `akkaPekko` and `fs2` modules.
 
-Moreover, examples contain two simple applications comprising Akka Streams or FS2 and Kafka. They show how you can write partitioned Parquet files with data coming from an indefinite stream.
+Moreover, examples contain two simple applications comprising Akka Streams / Pekko Streams or FS2 and Kafka. They show how you can write partitioned Parquet files with data coming from an indefinite stream.
