+++ site/docs/docs/partitioning.md
@@ -6,9 +6,9 @@ permalink: docs/partitioning/
 
 # Partitioning
 
-Parquet4s supports both reading partitions and partitioning data during writing. Writing partitioned data is available **only** in Akka and FS2 modules. Reading partitions is enabled by default in all Parquet4S modules.
+Parquet4s supports both reading partitions and partitioning data during writing. Writing partitioned data is available **only** in Akka / Pekko and FS2 modules. Reading partitions is enabled by default in all Parquet4S modules.
 
-#### Akka & FS2
+#### Akka / Pekko & FS2
 
 Reading partitions is handled by default by `fromParquet` function. Before data is read Parquet4s scans the directory and resolves partition fields and values. After reading, each record is enriched according to the partition directory tree the file resides in.
 
@@ -18,14 +18,15 @@ Writing partitioned data is available in `viaParquet`. You can specify by which 
 
 Since version 2.12.0 Parquet4S always reads partitioned data. Writing partitions is not supported in core module.
 
-#### **Take note!** 
+#### **Take note!**
 
- - The partition field must be a String. 
- - The field cannot be null, be an Option or belong to the collection.
- - When reading partitioned data make sure that partition directory names follow Hive format.
- - Parquet4s takes care of building proper schemas for partitioned data. However, when you use a custom type and a custom schema definition remember not to include the partition field in the schema — because it is supposed to be encoded as a directory name.
+- The partition field must be a String.
+- The field cannot be null, be an Option or belong to the collection.
+- When reading partitioned data make sure that partition directory names follow Hive format.
+- Parquet4s takes care of building proper schemas for partitioned data. However, when you use a custom type and a custom schema definition remember not to include the partition field in the schema — because it is supposed to be encoded as a directory name.
 
 In Akka:
+
 ```scala mdoc:compile-only
 import akka.NotUsed
 import akka.actor.ActorSystem
@@ -33,6 +34,48 @@ import akka.stream.scaladsl.{Sink, Source}
 import com.github.mjakubowski84.parquet4s.{Col, ParquetStreams, Path}
 
 object AkkaExample extends App {
+
+  implicit val actorSystem: ActorSystem = ActorSystem()
+  import actorSystem.dispatcher
+
+  case class PartitionDate(year: String, month: String, day: String)
+  case class User(id: Long, name: String, date: PartitionDate)
+
+  val users: Source[User, NotUsed] = ???
+  val path = Path("path/to/user/directory")
+
+  def writePartitionedData = users.via(
+    ParquetStreams
+      .viaParquet
+      .of[User]
+      .partitionBy(Col("date.year"), Col("date.month"), Col("date.day"))
+      .write(path)
+  ).runWith(Sink.foreach(user => println(s"Just wrote $user")))
+
+  def readPartitionedData = ParquetStreams
+    .fromParquet
+    .as[User]
+    .read(path)
+    .runWith(Sink.foreach(user => println(s"Just read $user")))
+
+  for {
+    _ <- writePartitionedData
+    _ <- readPartitionedData
+    _ <- actorSystem.terminate()
+  } yield ()
+
+}
+```
+
+In Pekko:
+
+```scala mdoc:compile-only
+import org.apache.pekko.NotUsed
+import org.apache.pekko.actor.ActorSystem
+import org.apache.pekko.stream.scaladsl.{Sink, Source}
+import com.github.mjakubowski84.parquet4s.{Col, ParquetStreams, Path}
+
+object PekkoExample extends App {
 
   implicit val actorSystem: ActorSystem = ActorSystem()
   import actorSystem.dispatcher
